Moved away from extending the ControllerBase class and created helper classes placed in the Fields folder, Fixed the file name and yml structure of the libraries.yml file, moved css folder and file into the module root, added code in .module to attach the library and css because it was originally not loading on the data dictionary form.
