+++ modules/data_dictionary_widget/css/dataDictionaryWidget.css
@@ -27,7 +27,7 @@
   height: 3rem;
   padding: 0.5rem 1rem;
   text-align: left; /* LTR */
-  color: #232429;
+  color: #292323;
   background: #f3f4f9;
   line-height: 1.25rem; /* 20px */
 }
