+++ modules/data_dictionary_widget/data_dictionary_widget.libraries.yml
@@ -1,0 +1,5 @@
+dataDictionaryWidget:
+  version: 1.x
+  css:
+    theme:
+      css/dataDictionaryWidget.css: {}
