+++ modules/data_dictionary_widget/data_dictionary_widget.module
@@ -6,7 +6,7 @@
  */
 
 use Drupal\Core\Entity\Display\EntityFormDisplayInterface;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldOperations;
+use Drupal\data_dictionary_widget\Fields\FieldOperations;
 use Drupal\node\NodeInterface;
 
 /**
@@ -67,8 +67,20 @@ function data_dictionary_widget__data_dictionary_data_type_checker($context) {
  * Setting form validation for unique identifier.
  * 
  * Modifying current_fields array structure to prevent errors in element render array.
+ * 
+ * Attaching module library to render theme changes.
  */
 function data_dictionary_widget_form_alter(&$form, &$form_state, $form_id) {
+  $formObject = $form_state->getFormObject();
+  if ($formObject instanceof \Drupal\Core\Entity\EntityFormInterface) {
+    $entity = $formObject->getEntity();
+    if ($form["field_json_metadata"]["widget"][0]["dictionary_fields"]) {
+      if ($entity->getEntityTypeId() === 'node' && in_array($entity->bundle(), ['data'])) {
+        $form['#attached']['library'][] = 'data_dictionary_widget/dataDictionaryWidget';
+      }
+    }
+  }
+
   if ($form_id == 'node_data_edit_form' || $form_id == 'node_data_form') {
     $form['#validate'][] = 'data_dictionary_widget_validate_unique_identifier';
     $current_fields = !empty($form["field_json_metadata"]["widget"][0]["dictionary_fields"]["current_fields"]) ? $form["field_json_metadata"]["widget"][0]["dictionary_fields"]["current_fields"] : NULL;
