--- modules/data_dictionary_widget/sdis_dashboards.libraries.yml
@@ -1,4 +1,0 @@
-data_dictionary_widget.style:
-  css:
-    layout:
-     src\css\dataDictionaryWidget.css: {}
