+++ modules/data_dictionary_widget/tests/src/Unit/DataDictionaryWidgetBuildTest.php
@@ -8,13 +8,13 @@ use Drupal\Core\Form\FormState;
 use Drupal\Core\Logger\LoggerChannelFactory;
 use Drupal\Core\StringTranslation\TranslationManager;
 use Drupal\metastore\SchemaRetriever;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldAddCreation;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldButtons;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldCallbacks;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldCreation;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldEditCreation;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldOperations;
-use Drupal\data_dictionary_widget\Controller\Widget\FieldValues;
+use Drupal\data_dictionary_widget\Fields\FieldAddCreation;
+use Drupal\data_dictionary_widget\Fields\FieldButtons;
+use Drupal\data_dictionary_widget\Fields\FieldCallbacks;
+use Drupal\data_dictionary_widget\Fields\FieldCreation;
+use Drupal\data_dictionary_widget\Fields\FieldEditCreation;
+use Drupal\data_dictionary_widget\Fields\FieldOperations;
+use Drupal\data_dictionary_widget\Fields\FieldValues;
 use MockChain\Options;
 use Drupal\Core\Field\WidgetBase;
 use Drupal\Core\Controller\ControllerBase;
